src/requirements.txt
annotated-types==0.5.0
anyio==3.7.1        
azure-common==1.1.28
azure-core==1.28.0  
azure-identity==1.13.0
azure-mgmt-core==1.4.0
azure-mgmt-desktopvirtualization==1.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
cryptography==41.0.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.7.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.7.0
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.23.1